\name{MAP.SIZE.TOLERANCE}
\alias{MAP.SIZE.TOLERANCE}
\title{package constant: MAP.SIZE.TOLERANCE}
\usage{
MAP.SIZE.TOLERANCE
}
\description{
    This variable defines how big a difference clusthaplo tolerates when realigning
    the genetic maps of the haplotypes and the MCQTL consensus.
}
\examples{
    print(MAP.SIZE.TOLERANCE)
}
