\name{set.debug}
\alias{set.debug}
\title{Debugging clusthaplo}
\usage{
    set.debug(dbg)
}
\arguments{
    \item{dbg}{logical.}
}
\description{
    \code{set.debug(TRUE)} activates various debug messages. \code{set.debug(FALSE)} deactivates them.
    It is strongly recommended to NOT use \code{set.debug(TRUE)}, for it is quite pointless to the user.
}
